Update styles, refactor
