+++ x-pack/plugins/security_solution/public/detection_engine/rule_management/components/rule_details/constants.ts
@@ -5,4 +5,5 @@
  * 2.0.
  */
 
-export const DESCRIPTION_LIST_COLUMN_WIDTHS: [string, string] = ['50%', '50%'];
+export const DEFAULT_DESCRIPTION_LIST_COLUMN_WIDTHS: [string, string] = ['50%', '50%'];
+export const LARGE_DESCRIPTION_LIST_COLUMN_WIDTHS: [string, string] = ['30%', '70%'];
