+++ x-pack/plugins/security_solution/public/detection_engine/rule_management_ui/components/rules_table/upgrade_prebuilt_rules_table/upgrade_prebuilt_rules_table_context.tsx
@@ -8,12 +8,16 @@
 import type { Dispatch, SetStateAction } from 'react';
 import React, { createContext, useCallback, useContext, useMemo, useState } from 'react';
 import { EuiButton } from '@elastic/eui';
+import type { EuiTabbedContentTab } from '@elastic/eui';
 import { useIsUpgradingSecurityPackages } from '../../../../rule_management/logic/use_upgrade_security_packages';
 import { useInstalledSecurityJobs } from '../../../../../common/components/ml/hooks/use_installed_security_jobs';
 import { useBoolState } from '../../../../../common/hooks/use_bool_state';
 import { affectedJobIds } from '../../../../../detections/components/callouts/ml_job_compatibility_callout/affected_job_ids';
 import type { RuleUpgradeInfoForReview } from '../../../../../../common/api/detection_engine/prebuilt_rules';
-import type { RuleSignatureId } from '../../../../../../common/api/detection_engine/model/rule_schema';
+import type {
+  RuleSignatureId,
+  RuleResponse,
+} from '../../../../../../common/api/detection_engine/model/rule_schema';
 import { invariant } from '../../../../../../common/utils/invariant';
 import {
   usePerformUpgradeAllRules,
@@ -28,8 +32,9 @@ import {
   RuleDetailsFlyout,
   TabContentPadding,
 } from '../../../../rule_management/components/rule_details/rule_details_flyout';
-import { RuleDiffTab } from '../../../../rule_management/components/rule_details/json_diff/rule_diff_tab';
+import { RuleDiffTab } from '../../../../rule_management/components/rule_details/rule_diff_tab';
 import { MlJobUpgradeModal } from '../../../../../detections/components/modals/ml_job_upgrade_modal';
+import { useIsExperimentalFeatureEnabled } from '../../../../../common/hooks/use_experimental_features';
 import * as ruleDetailsI18n from '../../../../rule_management/components/rule_details/translations';
 import * as i18n from './translations';
 
@@ -114,6 +119,10 @@ export const UpgradePrebuiltRulesTableContextProvider = ({
     filter: '',
     tags: [],
   });
+
+  const isJsonPrebuiltRulesDiffingEnabled = useIsExperimentalFeatureEnabled(
+    'jsonPrebuiltRulesDiffingEnabled'
+  );
 
   const isUpgradingSecurityPackages = useIsUpgradingSecurityPackages();
 
@@ -261,6 +270,29 @@ export const UpgradePrebuiltRulesTableContextProvider = ({
     actions,
   ]);
 
+  const getRuleTabs = useCallback(
+    (rule: RuleResponse, defaultTabs: EuiTabbedContentTab[]): EuiTabbedContentTab[] => {
+      const activeRule = filteredRules.find(({ id }) => id === rule.id);
+
+      if (!activeRule) {
+        return defaultTabs;
+      }
+
+      const diffTab = {
+        id: 'updates',
+        name: ruleDetailsI18n.UPDATES_TAB_LABEL,
+        content: (
+          <TabContentPadding>
+            <RuleDiffTab oldRule={activeRule.current_rule} newRule={activeRule.target_rule} />
+          </TabContentPadding>
+        ),
+      };
+
+      return [diffTab, ...defaultTabs];
+    },
+    [filteredRules]
+  );
+
   return (
     <UpgradePrebuiltRulesTableContext.Provider value={providerValue}>
       <>
@@ -275,6 +307,7 @@ export const UpgradePrebuiltRulesTableContextProvider = ({
         {previewedRule && (
           <RuleDetailsFlyout
             rule={previewedRule}
+            size={isJsonPrebuiltRulesDiffingEnabled ? 'l' : 'm'}
             dataTestSubj="updatePrebuiltRulePreview"
             closeFlyout={closeRulePreview}
             ruleActions={
@@ -290,28 +323,7 @@ export const UpgradePrebuiltRulesTableContextProvider = ({
                 {i18n.UPDATE_BUTTON_LABEL}
               </EuiButton>
             }
-            getRuleTabs={(rule, defaultTabs) => {
-              const activeRule = filteredRules.find(({ id }) => id === rule.id);
-
-              if (!activeRule) {
-                return defaultTabs;
-              }
-
-              const diffTab = {
-                id: 'updates',
-                name: ruleDetailsI18n.UPDATES_TAB_LABEL,
-                content: (
-                  <TabContentPadding>
-                    <RuleDiffTab
-                      oldRule={activeRule.current_rule}
-                      newRule={activeRule.target_rule}
-                    />
-                  </TabContentPadding>
-                ),
-              };
-
-              return [diffTab, ...defaultTabs];
-            }}
+            getRuleTabs={isJsonPrebuiltRulesDiffingEnabled ? getRuleTabs : undefined}
           />
         )}
       </>
